\name{draw.effect}
\alias{draw.effect}
\title{Create nice effect plots for MLMs}
\description{}
\usage{
  draw.effect(model, effect.name, axes=NULL, labels=NULL, plot.colors=get.color(), save.dir, file.name, img.format="png", img.w=10, img.unit="cm", img.dpi=600, save.plot=TRUE, return.plot=FALSE)
}
\arguments{
  \item{model}{}
  \item{effect.name}{}
  \item{axes}{}
  \item{labels}{}
  \item{plot.colors}{}
  \item{save.dir}{}
  \item{file.name}{}
  \item{img.format}{}
  \item{img.w}{}
  \item{img.unit}{}
  \item{img.dpi}{}
  \item{save.plot}{}
  \item{return.plot}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}