\name{draw.overlay}
\alias{draw.overlay}
\title{Create brain figure with overlays, legend, and key slice}
\description{}
\usage{
  draw.overlay(anat.nii, over.nii, over.vol, over.color, mask.nii, mask.vol, roi.nii, roi.val, roi.color="#ff64ff",orientation, idx.slice=TRUE, cbars=TRUE, save.dir, file.name, img.format="png", img.w=NULL, img.unit="in", img.dpi=600, save.plot=TRUE, return.plot=FALSE)
}
\arguments{
  \item{anat.nii}{}
  \item{over.nii}{}
  \item{over.vol}{}
  \item{over.color}{}
  \item{mask.nii}{}
  \item{mask.vol}{}
  \item{roi.nii}{}
  \item{roi.val}{}
  \item{roi.color}{}
  \item{orientation}{}
  \item{idx.slice}{}
  \item{cbars}{}
  \item{save.dir}{}
  \item{file.name}{}
  \item{img.format}{}
  \item{img.w}{}
  \item{img.unit}{}
  \item{img.dpi}{}
  \item{save.plot}{}
  \item{return.plot}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}