\name{slices.to.raster}
\alias{slices.to.raster}
\title{Convert 3d array of images to 2d matrix of those images}
\description{}
\usage{
  slices.to.raster(in.array, n.row, n.col, orientation)
}
\arguments{
  \item{in.array}{}
  \item{n.row}{}
  \item{n.col}{}
  \item{orientation}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}